backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.1
python-dotenv==1.0.0
pydantic==2.8.2
pydantic-settings==2.3.4
aiosqlite==0.20.0
aiofiles==24.1.0
httpx==0.27.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
Pillow==10.4.0
redis==5.0.1
opencv-python==4.9.0.80
opencv-contrib-python==4.9.0.80
ffmpeg-python==0.2.0
numpy==1.26.4
reportlab>=4.0.0
langdetect>=1.0.9

# Video Generation
moviepy>=1.0.3
gtts>=2.3.2
edge-tts>=6.1.0
pydub>=0.25.1
deep-translator>=1.11.4
bcrypt==4.0.1
openai>=1.0.0
